refactor `concurrent_imports` support in wasmtime-wit-bindgen
I had a few goals with this PR:

  1. Improve the ergonomics of concurrent import bindings by supporting
  `async`/`await` sugar and allowing store access to be arbitrarily interspersed
  between `await` points -- while still preventing references to the store
  across `await` points.

  2. Get rid of the `Data` associated types for `Host` traits and support
  `add_to_linker_get_host` where the `Host` impl is not the same as the `T` in
  `Store<T>`.

  3. Allow creating, reading from, writing to, etc. streams and futures without
  exposing `StoreContextMut` directly.

Unfortunately, after a day of intense type tetris I failed to achieve items 2 or
3, so this only covers item 1.

Regarding item 1: I've introduced a new `Accessor` type which wraps a `*mut dyn
VMStore` and provides access to it only via a `with` method that accepts a
synchronous closure which takes a `StoreContextMut<T>` parameter.  The closure
can do what it likes and return an arbitrary value as long as that result has a
`'static` lifetime (i.e. does not borrow from the store).  This ensures that the
host function is able to access the store only between `await`s and not across
them; we prohibit the latter because it would prevent other async-lowered
imports from running concurrently.  Finally, since host function takes a `&mut
Accessor<T>`, it is not possible for the reference to outlive the future
returned by the host function, and since the `with` method takes `&mut self` it
cannot be used recursively.

Regarding items 2 and 3: In order to read from or write to streams/futures, we
need to be able to efficiently lift and lower their payload types, which
requires that both the payload type (of which there could be several for a given
world) and the `T` in `Store<T>` be in scope.  I was unable to find a way to
thread those types through the various layers of closures, futures, and
generated code without adding unwanted `'static` bounds and/or breaking the
blanket `impl`s used for forwarding calls from `&mut X` to `X`.  Also, the usual
tricks of using dyn objects or vtables could only be used ergonomically to erase
one of the two types but not both.  I'd love to revisit this later with the help
of a Rust type wizard to see if there's a strategy I missed.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>