feat: update to independent trailer retrieval
This commit updates `wasi:http/types#body.finish` to avoid resolving
optional trailers, but instead returning them for the caller to
resolve.

This may introduce some efficiency gains in the cases where trailers
can be safely ignored, and does avoid the work done in checking
whether they are present or not, at least.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>