refactor `concurrent_imports` support in `wasmtime-wit-bindgen`
This is yet another major rework of `concurrent_imports`; this time, the goal is
to support `add_to_linker_get_host` properly such that input and output types of
the `host_getter` param need not be the same.  After much perserverance with
Alex and Lann, we arrived at something the compiler would accept, but it
requires thread-local storage and `unsafe` here and there, alas.

One drawback of this approach is that we no longer generate forwarding `impl`s
for `&mut T`, and thus no `add_to_linker` functions.  That's related to
concurrent functions taking `&mut Accessor<T, Self>` as a parameter and the
borrow checker not allowing us to map an `Accessor<T, &mut U>` to a `Accessor<T,
U>` despite all our efforts to convince it.

Finally, this adds `SpawnHandle` and `AbortOnDropHandle` types for aborting
`spawn`ed tasks.  I haven't actually tested this yet, but Roman has a test case
for his WASIp3 implementation, so I'm sure he'll let me know if there are
problems.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>