fix: avoid race condition in TCP stream handling drop
On some occasions `Notify::notified()` might have been called *after*
the `Notify::notify_waiters` call in the `drop`. Since `Notify` does
*not* record the notification, the task was never receiving it.

`ReceiveTask` does not actually need to know about the abort, as it can
infer the fact there is no more work to be done from the fact that the
sender side of the stream is closed.

It is assumed that send on a Future send handle, for which receiving end
is dropped it effectively a no-op, but, most importantly, it does not
block.

Signed-off-by: Roman Volosatovs <rvolosatovs@riseup.net>