refactor host stream/future APIs
This adds close-on-drop functionality to `{Stream|Future}{Writer|Reader}` and
support for reading and writing without a `StoreContextMut`.  Both of those
features are enabled by spawning background tasks which do the actual reading,
writing, and closing.

I've also tweaked the return types for `{Stream|Future}Reader` to reduce nesting
and reflect every possible outcome:
- Value(s) were read
- Write end closed with error before value(s) could be read
- Write end closed without error before value(s) could be read (streams only)
- Guest trapped before value(s) could be read (streams or futures)

Note that this introduces new `Host{Stream|Future}` types representing a stream
or future which has either just been lifted from the guest or is about to be
lowered to a guest.  These can be converted into `{Stream|Future}Reader`s via
the `into_reader` method, which requires a `StoreContextMut` in order to spawn
the background task mentioned above.  Ideally, we would support lifting and
lowering `{Stream|Future}Reader`s directly, in which case we could eliminate the
`Host{Stream|Future}` types, but we don't currently have access to a
`StoreContextMut` when lifting with a `LiftContext`, and it's not clear how to
work around that without some major refactoring.

Fixes #28

Signed-off-by: Joel Dice <joel.dice@fermyon.com>