add `watch_{reader|writer}` to `{Stream|Future}{Writer|Reader}` (#54)
These methods allow the caller to wait for the other end to be closed.  This can
be useful in the case where the owner of the write end needs to `await` input
from some other source before it can write, in which case it may want to select
over both the input future and the future returned by `watch_reader`, canceling
the former if the reader is closed.

Note that `watch_{reader|writer}` takes `self` by value and returns a
`Watch<Self>` which can be converted back into a `Self` at any time using
`Watch::into_inner`.  Continuing the example above: if the input future yields a
value before the read end is closed, then the writer can call
`Watch::into_inner` to get the `StreamWriter` back and call
`StreamWriter::write` on it, then loop around again to read more input.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>