simplify handling of `StoreContextMut` in concurrent.rs
Previously, several functions in concurrent.rs which accepted a
`StoreContextMut` parameter also returned a `StoreContextMut`.  That was
motivated by a concern that functions which (directly or transitively) convert a
`&mut StoreInner` into a raw pointer would invalidate the original reference.
However, that turned out to be unnecessary paranoia; it's no less sound than a
reborrow assuming we scope the use of that raw pointer appropriately.

This commit removes the unnecessary return values, which helps simplify both the
internal code and the `pub(crate)` API.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>