Enable async support in wast tests
This commit updates the `wasmtime-wast` crate to conditionally use the
`Config::async_support` mode of Wasmtime, notably executing
`instantiate_async` and `call_async`. At this time no actual concurrency
is supported and everything is immediately await'd on via `block_on` and
a local single-thread tokio runtime.

The motivation for this commit is that in the upcoming implementation of
WASIp3 async will effectively be the only way to invoke components.
Furthermore to test these components we're wanting to use async APIs as
the root of invocations for writing `*.wast` tests for testing
component-model async features.

In enabling this support this commit additionally updates the
`wast_tests` fuzzer to execute all tests both with and without async.
Effectively the async configuration is now a fuzz option, meaning that
all tests are being executed with async now as well to ensure that they
work (yay!).

The `--test wast` testing mode and `wasmtime wast` CLI command are both
updated to unconditionally use async. There should be no loss in test
coverage due to the fuzzer update, and right now knobs aren't provided
to conditionally use sync in these two modes, again because it's
currently expected that this won't be able to run component model async
tests, so the defaults are changing.