validate task.return canon options at runtime (#70)
* validate task.return canon options at runtime

Per https://github.com/bytecodealliance/wasmtime/issues/10338, we need to
validate three things match for every call to `task.return` relative to the
canonical lift options for the export being executed:

- The type being returned
- The memory option (if applicable)
- The string encoding (if applicable)

We were already checking the first of those; this adds the other two.

In the process of doing this, Alex noticed that we were interning imports
incorrectly in FACT.  I've disabled interning entirely in some cases where it's
not worth it, and I've fixed the remaining cases to use appropriate key types.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* add `task.return` sad path tests

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

---------

Signed-off-by: Joel Dice <joel.dice@fermyon.com>