update CM async support to match spec (#79)
This updates the CM async support to match recent (and not-so-recent) spec
changes:

- Use waitable-sets to track and deliver events for waitables
- Expect packed (`CallbackCode`, waitable-set) tuples from exports and callbacks
- Remove `Event.DONE`, which no longer exists
- Update the semantics of `subtask.drop` given that it can be called before the subtask has finished
- Allow the read and write ends of a stream or future to be added to separate waitable-sets

The last item required a fair amount of refactoring to give each end of a stream
or future its own handle.

The `subtask.drop`/`Event.DONE` changes mean that the parent of a subtask may
finish before the subtask does, in which case we must "reparent" the subtask to
point to its grandparent (or host, if applicable).  This is akin to an "async
tail call" such that long-lived subtasks "bubble up" the call stack as its
caller (and it's caller's caller, etc.) finish.

Note that I've temporarily pointed the wit-bindgen and wasm-tools deps to my
fork; I'll open PRs to get those changes upstreamed soon.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>