various fixes related to stream/future cancellation (#89)
Most of this is related to the delay between queuing an event (e.g. read
completion) and delivering it; a guest could cancel inbetween, and we should
allow that rather than trap.

Also, we need to be careful to invalidate any queued events as part of
cancellation, especially given that the guest could cancel and close a stream or
future before delivery.

This code is getting uglier with time and will likely need to be cleaned up and
refactored before it is upstreamed.

Fixes #84

Signed-off-by: Joel Dice <joel.dice@fermyon.com>