add `Instance::promise` for wrapping a `Future` in a `Promise`
This may be useful as e.g. an alternative or complement to `PromisesUnordered`
where you want to compose some number of `Promise`s in an `async` function or
block with arbitrary control flow that might be cumbersome to express using
`PromisesUnordered` by itself.  It can also be used to reverse a previous
`Promise::into_future` operation.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>