add `Instance::promise` for wrapping a `Future` in a  `Promise` (#98)
* use `StreamWriter::write_all` where appropriate in tests

This probably isn't necessary for tests which only write one item at a time
anyway, but it at least communicates the intent better.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* add `Instance::promise` for wrapping a `Future` in a `Promise`

This may be useful as e.g. an alternative or complement to `PromisesUnordered`
where you want to compose some number of `Promise`s in an `async` function or
block with arbitrary control flow that might be cumbersome to express using
`PromisesUnordered` by itself.  It can also be used to reverse a previous
`Promise::into_future` operation.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

---------

Signed-off-by: Joel Dice <joel.dice@fermyon.com>