Update WIT in tests with `async`
This commit updates the WIT used in test to use the `async` keyword and
generally try to avoid manually configuring what's async and what's not
in the guest. This does not update the host's handling of the `async`
keyword in WIT for `wasmtime::component::bindgen!`. Some assorted
bindgen issues were fixed here and there and WASI WITs were updated as
well with `async` where it seemed appropriate. I'll note that there are
some usages of concurrent imports when there's no actual concurrency,
it's just there for getting an `Accessor` and, for example, spawning
tasks. In these situations we may want to explore possible modes of
bindings generation on the host side of things which avoids the need for
manual configuration.

Coupled with bytecodealliance/wit-bindgen#1278 and
bytecodealliance/wit-bindgen#1279.