refactor/rewrite `concurrent.rs` event loop
The event loop and scheduler in `concurrent.rs` have undergone a lot of changes
over time, reflecting significant changes in the component model spec.  Some
features, such as yielding, backpressure, and waitable sets, were retrofitted
onto the existing code in a somewhat haphazard way, making the code more
convoluted than it needed to be.  In addition, it made control flow hard to
understand and debug, with multiply-nested fibers and no clear division of
responsibility among tasks and the fibers they ran on.

This commit rewrites most of that code to make it more cohesive:

- No nested fibers: we always suspend the current fiber before resuming the next one
- One single function (`poll_until`) runs the event loop (rather than several functions with slightly different semantics, as before)
- The code paths for `{Typed}Func::call_{async,concurrent}` have been unified, reducing redundancy and unsafe code
- We now manage reentrance, yielding, and backpressure by explicitly deferring work as appropriate
    - Work is deferred to either:
        - a high priority queue (work to be done ASAP)
        - a low priority queue (work to be done only once the high priority queue has been drained)
        - a per-instance `pending` bucket (work that can't be done until a current activation for that instance returns and/or backpressure is unset)
- The new code is more `waitable-set`-native, removing obsolete ties between parent and child tasks which no longer apply
- Fixed previous unsoundness due to recursively modifying `ConcurrentState::futures` while polling it
- More types implement `Debug`, making logs easier to follow

Signed-off-by: Joel Dice <joel.dice@fermyon.com>