Publish `dev` artifacts in CI
In thinking about how to set up `wit-bindgen` tests I'm reversing course
and thinking it might be better to have everything live in
`wit-bindgen`. Otherwise figuring out how to get the patches and stars
to align to get everything tested feels kind of daunting, so my hope is
that we'll eat some instability in the `wit-bindgen` CI for now by
downloading a `dev` version of the `wasmtime` binary from here for async
tests. If that causes too much instability in the `wit-bindgen` CI I
figure we can reevaluate at that point.