change the error returned by `Instance::run` etc. on deadlock
Previously, we returned an "async-lifted export failed to return a result" which
was not always accurate and usually misleading.

This also adds documentation to `Instance::run` explaining when and why a
"deadlock" error will be returned.

Note that the new `backpressure-deadlock.wast` test was copied from Alex
Crichton's issue description.

Fixes #152

Signed-off-by: Joel Dice <joel.dice@fermyon.com>