fix memory leak when panicking with detached `ComponentInstance`
In `ComponentInstance::from_vmctx` and
`StoreContextMut::with_detached_instance[_async]` we were leaking memory due to
having taken `InstanceData` out of the `Store` and making it unreachable except
via a raw pointer, meaning there was nothing responsible for dropping it on
panic.  This commit stashes the data elsewhere in the store to make sure it gets
dropped.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>