Migrate http tests to `wasmtime-wasi-http`
This removes the `wasi-http-draft` in this repository as well as the
draft set of `*.wit` files for a historical version of wasi-http.
Existing tests are migrated to the `wasmtime-wasi-http` crate using
various utilities and helpers there instead. This is done to cut down on
the amount of code to update and to additionally exercise the more of
the `wasmtime-wasi-http` crate as well.