eliminate `ComponentInstance` unsafe code in futures_and_streams.rs
This is the first step in eliminating the need for `*mut ComponentInstance` in
most or all cases, instead preferring `Instance` indexes.

The next step will be do remove the
`StoreContextMut::with_{a,de}tached_instance[_async]` functions and update any
function signatures which take both `&mut dyn VMStore` (or similar) and `&mut
ComponentInstance` parameters and replace the latter with an `Instance`
parameter.  For the time being, such functions are _extremely_ hazardous, since
it's easy to alias the `&mut ComponentInstance` by way of the store.  So for
now, don't do that!

Signed-off-by: Joel Dice <joel.dice@fermyon.com>