Remove no-longer-needed helper method
In trying to upstream this helper method it had surprising semantics for
insert-beyond-the-end and so upon investigating I concluded that this
may be a historical artifact at this point and no longer needed. To
reduce the diff with upstream this commit removes the method and
replaces the one usage with a locally-equivalent structure.