remove `StoreContextMut::with_{a,de}tached_instance[_async]`
This is the second phase of the cleanup I started in an earlier commit with the
goal of reducing `unsafe` code in the `concurrent` module and submodules,
especially with regard to functions which take mutable references to a store and
a `ComponentInstance` within that store.

My earlier attempt to make such functions safe was to temporarily "detach" the
instance from the store to avoid aliasing hazards, but that got messy and
confusing, and Alex and I decided an index-based approach made more sense.  So
now we pass `Instance` handles around instead of `ComponentInstance` references
and pointers, using them to retrieve the `ComponentInstance` from the store only
as necessary.

Practically speaking, this required moving a lot of inherent functions from
`ComponentInstance` to `Instance` and updating their bodies accordingly.
Functionally, nothing has changed.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>