Simplify thread-local management of the store
This commit is a simplification of the management of thread locals used
to transfer a store across the `Future::poll` boundary to pass it from
the caller to internal futures we're polling. Previously there were two
thread locals where a store would migrate between them as appropriate,
and these two thread locals were adjacent to the rest of `concurrent.rs`
which accessed it from a number of locations. After this commit all TLS
accesses are isolated in a single `tls.rs` file which is intended to be
managable to audit in its entirety.

Both TLS globals have been folded in a single global that is two
pointers large, storing a `dyn VMStore` pointer. The store now has
internal state associated with the current instance being polled as well
which removes th need for some extra state to be passed in these
globals. Additionally functions such as `with_local_instance` where two
arguments were produced are now split into one argument (the store)
comes from TLS and the other (the instance) comes from a closed over
argument in futures/etc. The culmination of this is that all TLS
management goes through a smaller set of primitives than before which is
intended to be easier to audit for unsafe.