implement new `stream`/`future` trap conditions
As of https://github.com/WebAssembly/component-model/pull/524, we now trap if a
guest tries to use a stream or future handle (besides dropping it) after it is
"done" (see that PR for what "done" means).  This required more careful tracking
of the state of each end of the stream or future.

In the process of debugging this into shape, I added some trace logging and
addressed a couple of issues I noticed:

- `wasi-http`'s `Response::into_http` wasn't noticing a non-empty buffer at the end of a stream
- the `p3_http_middleware` test was not clearing its buffer before each read

Signed-off-by: Joel Dice <joel.dice@fermyon.com>