combine fiber abstractions from `async_.rs` and `concurrent.rs`
A long time ago, when I first started implementing CM async, I cut-and-pasted
`async_.rs`'s `FiberFuture` as `StoreFiber` and heavily modified it.  Both have
evolved separately since then, and now that the CM async work has generally
stabilized it's time to unify them.

This creates a new `runtime::fiber` submodule which `async_.rs` and
`concurrent.rs` share.  The central type is `StoreFiber`, which is a more
versatile version of `FiberFuture`, supporting fibers which can either retain
exclusive access to the store across suspend points or release it, allowing the
store to be used by the root event loop and/or other fibers to run before the
original fiber resumes.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>