remove `TcpStream::set_nodelay` call
This is no longer necessary to avoid 40ms of extra response latency for
`wasi:http@0.3.0-draft` components since we now poll the body stream before
handing the response to Hyper.  I still think the Nagle algorithm is a bad
default, but I can make a separate PR to disable it.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>